project/backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
aiomysql>=0.2.0
pymysql>=1.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Email
aiosmtplib>=3.0.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Utilities
python-dotenv>=1.0.0
APScheduler>=3.10.0
pytz>=2023.3
redis>=5.0.0
